process_packets now only processes a given command
